updated still not working

+++ lib/features/rentee/renting_status/presentation/widgets/inrenting_item_card_widget.dart
@@ -73,7 +73,7 @@ class _InrentingItemCardWidgetState extends State<InrentingItemCardWidget> {
           (context) => RenteeQRDialog(
             item: widget.item,
             orderDate: rentalDays,
-            totalFee: widget.totalPrice.toInt(),
+            totalFee: widget.totalPrice,
             currentStatus: widget.status,
             onSimulateScan: () => _simulateReturnScan(context),
             showSimulateButton: kIsWeb,
